Progwards: bitsworld(2)

+++ src/ru/progwards/java1/lessons/bitsworld/Binary.java
@@ -3,9 +3,9 @@ package ru.progwards.java1.lessons.bitsworld;
 import org.junit.Assert;
 
 public class Binary {
-    private byte num;
+    private int num;
 
-    public Binary(byte num) {
+    public Binary(int num) {
         this.num = num;
     }
 
@@ -19,7 +19,7 @@ public class Binary {
     }
 
     public static void main(String[] args) {
-        Assert.assertEquals(new Binary((byte) 0b00001101).toString(), "00001101");
-        Assert.assertEquals(new Binary((byte) 0b01010101).toString(), "01010101");
+        Assert.assertEquals(new Binary(0b00001101).toString(), "00001101");
+        Assert.assertEquals(new Binary(0b01010101).toString(), "01010101");
     }
 }
